Adds IsFirstByMonitorIdAsync and Improves Event Logic
The IsFirstByMonitorIdAsync method has been added to IEventRepository and implemented in EventRepository to check if there are previous events for a monitor. In MonitorJob, the important event logic has been modified to also consider whether it is the monitor's first event. Additionally, unnecessary usings have been removed from EventRepository.cs.