Refactors GenerateMonitorReport and improves Excel formatting
The signature of GenerateMonitorReport in IReportService and ReportService has been changed to use DateTime instead of string for the dateFrom and dateTo parameters. In GenerateMonitorReportCommandHandler, the call to pass DateTime objects has been adjusted. In ReportService, the date range in Excel is now displayed in dd/MM/yyyy format. The calculation and writing of the total time difference between dateFrom and dateTo in cell (9,7) has been added. DownTime is now displayed in hh:mm:ss format instead of hh:mm.