ComponentMonitor's IsUp now allows null values
The IsUp property of the ComponentMonitor entity has been changed to nullable (bool?), allowing null values ​​in addition to true/false. The model configuration and context snapshot have been updated to reflect this change, removing the IsRequired() requirement. Migration files have been added that modify the IsUp column in the database to accept null values ​​and define the rollback of the change, restoring the constraint and the default value.